refactor: move audit log handler to auditlog sub-package
