+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/handlers/auditlog/AuditLogHandler.java
@@ -5,12 +5,13 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.exporter.handlers;
+package io.camunda.exporter.handlers.auditlog;
 
 import static java.util.Optional.ofNullable;
 
 import io.camunda.exporter.exceptions.PersistenceException;
-import io.camunda.exporter.handlers.AuditLogHandler.AuditLogBatch;
+import io.camunda.exporter.handlers.ExportHandler;
+import io.camunda.exporter.handlers.auditlog.AuditLogHandler.AuditLogBatch;
 import io.camunda.exporter.store.BatchRequest;
 import io.camunda.webapps.schema.descriptors.template.AuditLogTemplate;
 import io.camunda.webapps.schema.entities.ExporterEntity;
